Remove SVGLocatable and SVGTransformable https://bugs.webkit.org/show_bug.cgi?id=250797 rdar://104668934
Reviewed by Nikolas Zimmermann.

Per the SVG2 spec changes (https://w3c.github.io/svgwg/svg2-draft/changes.html#types),
SVGLocatable and SVGTransformable have been folded into SVGGraphicsElement,
which directly inherits from SVGElement.

This patch removes SVGLocatable.{cpp,h}, SVGTransformable.{cpp,h}, and
SVGTransformableInlines.h, moving their functionality into
SVGGraphicsElement and SVGTransformList. It also addresses a Safer C++
unchecked call argument issue in SVGGraphicsElement::computeBBox by
storing the renderer in a CheckedPtr local.

* Source/WebCore/SaferCPPExpectations/UncheckedCallArgsCheckerExpectations:
* Source/WebCore/Sources.txt:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/loader/FrameLoader.cpp:
* Source/WebCore/page/InteractionRegion.cpp:
(WebCore::interactionRegionForRenderedRegion):
* Source/WebCore/rendering/RenderObject.h:
* Source/WebCore/rendering/svg/RenderSVGModelObject.cpp:
(WebCore::RenderSVGModelObject::checkIntersection):
(WebCore::RenderSVGModelObject::checkEnclosure):
* Source/WebCore/rendering/svg/RenderSVGShape.cpp:
(WebCore::RenderSVGShape::nonScalingStrokeTransform const):
* Source/WebCore/rendering/svg/legacy/LegacyRenderSVGModelObject.cpp:
(WebCore::getElementCTM):
* Source/WebCore/rendering/svg/legacy/LegacyRenderSVGResourceContainer.cpp:
(WebCore::LegacyRenderSVGResourceContainer::transformOnNonScalingStroke):
* Source/WebCore/rendering/svg/legacy/LegacyRenderSVGShape.cpp:
(WebCore::LegacyRenderSVGShape::nonScalingStrokeTransform const):
* Source/WebCore/svg/SVGAnimateTransformElement.cpp:
(WebCore::SVGAnimateTransformElement::attributeChanged):
* Source/WebCore/svg/SVGGradientElement.cpp:
* Source/WebCore/svg/SVGGraphicsElement.cpp:
(WebCore::isViewportElement):
(WebCore::SVGGraphicsElement::nearestViewportElement):
(WebCore::SVGGraphicsElement::computeBBox):
(WebCore::SVGGraphicsElement::computeCTM):
(WebCore::SVGGraphicsElement::getCTM):
(WebCore::SVGGraphicsElement::getScreenCTM):
(WebCore::SVGGraphicsElement::getBBox):
* Source/WebCore/svg/SVGGraphicsElement.h:
* Source/WebCore/svg/SVGLocatable.cpp: Removed.
* Source/WebCore/svg/SVGLocatable.h: Removed.
* Source/WebCore/svg/SVGPatternElement.cpp:
* Source/WebCore/svg/SVGSVGElement.cpp:
(WebCore::SVGSVGElement::attributeChanged):
* Source/WebCore/svg/SVGTransformList.cpp:
(WebCore::parseTransformTypeGeneric):
(WebCore::SVGTransformList::parseTransformType):
(WebCore::SVGTransformList::parseGeneric):
* Source/WebCore/svg/SVGTransformList.h:
* Source/WebCore/svg/SVGTransformListInlines.h: Renamed from Source/WebCore/svg/SVGTransformableInlines.h.
(WebCore::SVGTransformList::parseAndReplaceTransform):
(WebCore::SVGTransformList::parseTransform):
* Source/WebCore/svg/SVGTransformable.cpp: Removed.
* Source/WebCore/svg/SVGTransformable.h: Removed.
* Source/WebCore/svg/SVGViewSpec.cpp:

Canonical link: https://commits.webkit.org/309396@main
