+++ Source/WebCore/SaferCPPExpectations/UncheckedCallArgsCheckerExpectations
@@ -605,7 +605,6 @@ svg/SVGFontFaceElement.cpp
 svg/SVGFontFaceUriElement.cpp
 svg/SVGImageElement.cpp
 svg/SVGLengthContext.cpp
-svg/SVGLocatable.cpp
 svg/SVGMaskElement.cpp
 svg/SVGSVGElement.cpp
 svg/SVGToOTFFontConversion.cpp
