+++ Source/WebCore/svg/SVGAnimateTransformElement.cpp
@@ -24,7 +24,7 @@
 #include "SVGAnimateTransformElement.h"
 
 #include "SVGNames.h"
-#include "SVGTransformable.h"
+#include "SVGTransformList.h"
 #include <wtf/TZoneMallocInlines.h>
 
 namespace WebCore {
@@ -57,7 +57,7 @@ bool SVGAnimateTransformElement::hasValidAttributeType() const
 void SVGAnimateTransformElement::attributeChanged(const QualifiedName& name, const AtomString& oldValue, const AtomString& newValue, AttributeModificationReason attributeModificationReason)
 {
     if (name == SVGNames::typeAttr) {
-        m_type = SVGTransformable::parseTransformType(newValue).value_or(SVGTransformValue::SVG_TRANSFORM_UNKNOWN);
+        m_type = SVGTransformList::parseTransformType(newValue).value_or(SVGTransformValue::SVG_TRANSFORM_UNKNOWN);
         if (m_type == SVGTransformValue::SVG_TRANSFORM_MATRIX)
             m_type = SVGTransformValue::SVG_TRANSFORM_UNKNOWN;
     }
