+++ Source/WebCore/svg/SVGGraphicsElement.cpp
@@ -1,7 +1,9 @@
 /*
  * Copyright (C) 2004, 2005, 2008 Nikolas Zimmermann <zimmermann@kde.org>
  * Copyright (C) 2004, 2005, 2006 Rob Buis <buis@kde.org>
+ * Copyright (C) 2009 Google, Inc. All rights reserved.
  * Copyright (C) 2018 Apple Inc. All rights reserved.
+ * Copyright (C) Research In Motion Limited 2010. All rights reserved.
  *
  * This library is free software; you can redistribute it and/or
  * modify it under the terms of the GNU Library General Public
@@ -26,13 +28,18 @@
 #include "LegacyRenderSVGPath.h"
 #include "LegacyRenderSVGResource.h"
 #include "RenderAncestorIterator.h"
+#include "RenderElement.h"
 #include "RenderElementStyleInlines.h"
 #include "RenderLayer.h"
 #include "RenderLayerInlines.h"
+#include "RenderObjectDocument.h"
 #include "RenderSVGHiddenContainer.h"
 #include "RenderSVGPath.h"
 #include "RenderSVGResourceMasker.h"
 #include "RenderSVGResourcePattern.h"
+#include "SVGElementTypeHelpers.h"
+#include "SVGImageElement.h"
+#include "SVGLayerTransformComputation.h"
 #include "SVGMatrix.h"
 #include "SVGNames.h"
 #include "SVGPathData.h"
@@ -40,14 +47,91 @@
 #include "SVGRenderSupport.h"
 #include "SVGSVGElement.h"
 #include "SVGStringList.h"
+#include "Settings.h"
 #include "StyleTransformResolver.h"
 #include "TransformOperationData.h"
+#include "TransformState.h"
 #include <wtf/NeverDestroyed.h>
 #include <wtf/TZoneMallocInlines.h>
 
 namespace WebCore {
 
 WTF_MAKE_TZONE_ALLOCATED_IMPL(SVGGraphicsElement);
+
+// FIXME: This doesn't match SVGElement::viewportElement() as it has an extra check for
+// foreign object.
+static bool NODELETE isViewportElement(const SVGElement* element)
+{
+    if (!element)
+        return false;
+
+    return element->hasTagName(SVGNames::svgTag)
+        || element->hasTagName(SVGNames::symbolTag)
+        || element->hasTagName(SVGNames::foreignObjectTag)
+        || is<SVGImageElement>(*element);
+}
+
+SVGElement* SVGGraphicsElement::nearestViewportElement(const SVGElement* element)
+{
+    ASSERT(element);
+    for (auto* current = element->parentOrShadowHostElement(); current; current = current->parentOrShadowHostElement()) {
+        auto* svgElement = dynamicDowncast<SVGElement>(*current);
+        if (isViewportElement(svgElement))
+            return svgElement;
+    }
+
+    return nullptr;
+}
+
+FloatRect SVGGraphicsElement::computeBBox(SVGElement* element, StyleUpdateStrategy styleUpdateStrategy)
+{
+    ASSERT(element);
+    if (styleUpdateStrategy == AllowStyleUpdate)
+        protect(element->document())->updateLayoutIgnorePendingStylesheets({ LayoutOptions::TreatContentVisibilityHiddenAsVisible, LayoutOptions::TreatContentVisibilityAutoAsVisible }, element);
+
+    // FIXME: Eventually we should support getBBox for detached elements.
+    CheckedPtr renderer = element->renderer();
+    if (!renderer || renderer->isRenderSVGHiddenContainer())
+        return FloatRect();
+
+    return renderer->objectBoundingBox();
+}
+
+AffineTransform SVGGraphicsElement::computeCTM(SVGElement* element, CTMScope mode, StyleUpdateStrategy styleUpdateStrategy)
+{
+    ASSERT(element);
+    if (styleUpdateStrategy == AllowStyleUpdate)
+        protect(element->document())->updateLayoutIgnorePendingStylesheets({ LayoutOptions::TreatContentVisibilityHiddenAsVisible, LayoutOptions::TreatContentVisibilityAutoAsVisible }, element);
+
+    RefPtr stopAtElement = mode == CTMScope::NearestViewportScope ? nearestViewportElement(element) : nullptr;
+
+    if (element->document().settings().layerBasedSVGEngineEnabled()) {
+        // Rudimentary support for operations on "detached" elements.
+        CheckedPtr renderer = dynamicDowncast<RenderLayerModelObject>(element->renderer());
+        if (!renderer)
+            return element->localCoordinateSpaceTransform(mode);
+
+        auto trackingMode { mode == CTMScope::ScreenScope ? TransformState::TrackSVGScreenCTMMatrix : TransformState::TrackSVGCTMMatrix };
+        CheckedPtr stopAtRenderer = dynamicDowncast<RenderLayerModelObject>(stopAtElement ? stopAtElement->renderer() : nullptr);
+        return SVGLayerTransformComputation(*renderer).computeAccumulatedTransform(stopAtRenderer.get(), trackingMode);
+    }
+
+    AffineTransform ctm;
+
+    for (RefPtr<Element> currentElement = element; currentElement; currentElement = currentElement->parentOrShadowHostElement()) {
+        RefPtr svgElement = dynamicDowncast<SVGElement>(*currentElement);
+        if (!svgElement)
+            break;
+
+        ctm = svgElement->localCoordinateSpaceTransform(mode).multiply(ctm);
+
+        // For getCTM() computation, stop at the nearest viewport element
+        if (currentElement == stopAtElement)
+            break;
+    }
+
+    return ctm;
+}
 
 SVGGraphicsElement::SVGGraphicsElement(const QualifiedName& tagName, Document& document, UniqueRef<SVGPropertyRegistry>&& propertyRegistry, OptionSet<TypeFlag> typeFlags)
     : SVGElement(tagName, document, WTF::move(propertyRegistry), typeFlags)
@@ -71,7 +155,7 @@ Ref<SVGMatrix> SVGGraphicsElement::getCTMForBindings()
 
 AffineTransform SVGGraphicsElement::getCTM(StyleUpdateStrategy styleUpdateStrategy)
 {
-    return SVGLocatable::computeCTM(this, CTMScope::NearestViewportScope, styleUpdateStrategy);
+    return computeCTM(this, CTMScope::NearestViewportScope, styleUpdateStrategy);
 }
 
 Ref<SVGMatrix> SVGGraphicsElement::getScreenCTMForBindings()
@@ -81,7 +165,7 @@ Ref<SVGMatrix> SVGGraphicsElement::getScreenCTMForBindings()
 
 AffineTransform SVGGraphicsElement::getScreenCTM(StyleUpdateStrategy styleUpdateStrategy)
 {
-    return SVGLocatable::computeCTM(this, CTMScope::ScreenScope, styleUpdateStrategy);
+    return computeCTM(this, CTMScope::ScreenScope, styleUpdateStrategy);
 }
 
 AffineTransform SVGGraphicsElement::animatedLocalTransform() const
@@ -168,7 +252,7 @@ Ref<SVGRect> SVGGraphicsElement::getBBoxForBindings()
 
 FloatRect SVGGraphicsElement::getBBox(StyleUpdateStrategy styleUpdateStrategy)
 {
-    return SVGTransformable::getBBox(this, styleUpdateStrategy);
+    return computeBBox(this, styleUpdateStrategy);
 }
 
 RenderPtr<RenderElement> SVGGraphicsElement::createElementRenderer(RenderStyle&& style, const RenderTreePosition&)
