+++ Source/WebCore/svg/SVGTransformList.cpp
@@ -1,4 +1,7 @@
 /*
+ * Copyright (C) 2004, 2005, 2008 Nikolas Zimmermann <zimmermann@kde.org>
+ * Copyright (C) 2004, 2005, 2006, 2007 Rob Buis <buis@kde.org>
+ * Copyright (C) 2007 Eric Seidel <eric@webkit.org>
  * Copyright (C) 2020 Apple Inc. All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -29,12 +32,61 @@
 #include "ExceptionOr.h"
 #include "SVGParserUtilities.h"
 #include "SVGTransform.h"
-#include "SVGTransformable.h"
-#include "SVGTransformableInlines.h"
+#include "SVGTransformListInlines.h"
 #include <wtf/text/StringBuilder.h>
 #include <wtf/text/StringParsingBuffer.h>
+#include <wtf/text/StringView.h>
 
 namespace WebCore {
+
+template<typename CharacterType> static constexpr std::array<CharacterType, 5> skewXDesc  { 's', 'k', 'e', 'w', 'X' };
+template<typename CharacterType> static constexpr std::array<CharacterType, 5> skewYDesc  { 's', 'k', 'e', 'w', 'Y' };
+template<typename CharacterType> static constexpr std::array<CharacterType, 5> scaleDesc  { 's', 'c', 'a', 'l', 'e' };
+template<typename CharacterType> static constexpr std::array<CharacterType, 9> translateDesc  { 't', 'r', 'a', 'n', 's', 'l', 'a', 't', 'e' };
+template<typename CharacterType> static constexpr std::array<CharacterType, 6> rotateDesc  { 'r', 'o', 't', 'a', 't', 'e' };
+template<typename CharacterType> static constexpr std::array<CharacterType, 6> matrixDesc  { 'm', 'a', 't', 'r', 'i', 'x' };
+
+template<typename CharacterType> static std::optional<SVGTransformValue::SVGTransformType> parseTransformTypeGeneric(StringParsingBuffer<CharacterType>& buffer)
+{
+    if (buffer.atEnd())
+        return std::nullopt;
+
+    if (*buffer == 's') {
+        if (skipCharactersExactly(buffer, std::span { skewXDesc<CharacterType> }))
+            return SVGTransformValue::SVG_TRANSFORM_SKEWX;
+        if (skipCharactersExactly(buffer, std::span { skewYDesc<CharacterType> }))
+            return SVGTransformValue::SVG_TRANSFORM_SKEWY;
+        if (skipCharactersExactly(buffer, std::span { scaleDesc<CharacterType> }))
+            return SVGTransformValue::SVG_TRANSFORM_SCALE;
+        return std::nullopt;
+    }
+
+    if (skipCharactersExactly(buffer, std::span { translateDesc<CharacterType> }))
+        return SVGTransformValue::SVG_TRANSFORM_TRANSLATE;
+    if (skipCharactersExactly(buffer, std::span { rotateDesc<CharacterType> }))
+        return SVGTransformValue::SVG_TRANSFORM_ROTATE;
+    if (skipCharactersExactly(buffer, std::span { matrixDesc<CharacterType> }))
+        return SVGTransformValue::SVG_TRANSFORM_MATRIX;
+
+    return std::nullopt;
+}
+
+std::optional<SVGTransformValue::SVGTransformType> SVGTransformList::parseTransformType(StringView string)
+{
+    return readCharactersForParsing(string, [](auto buffer) {
+        return parseTransformType(buffer);
+    });
+}
+
+std::optional<SVGTransformValue::SVGTransformType> SVGTransformList::parseTransformType(StringParsingBuffer<Latin1Character>& buffer)
+{
+    return parseTransformTypeGeneric(buffer);
+}
+
+std::optional<SVGTransformValue::SVGTransformType> SVGTransformList::parseTransformType(StringParsingBuffer<char16_t>& buffer)
+{
+    return parseTransformTypeGeneric(buffer);
+}
 
 ExceptionOr<RefPtr<SVGTransform>> SVGTransformList::consolidate()
 {
@@ -77,12 +129,12 @@ template<typename CharacterType> bool SVGTransformList::parseGeneric(StringParsi
     while (buffer.hasCharactersRemaining()) {
         delimParsed = false;
 
-        auto parsedTransformType = SVGTransformable::parseTransformType(buffer);
+        auto parsedTransformType = parseTransformType(buffer);
         if (!parsedTransformType)
             return false;
 
         if (currentListReplacement == ListReplacement::Replace && itemIndex < m_items.size() && parsedTransformType == m_items[itemIndex]->type()) {
-            if (!SVGTransformable::parseAndReplaceTransform(*parsedTransformType, buffer, m_items[itemIndex]))
+            if (!parseAndReplaceTransform(*parsedTransformType, buffer, m_items[itemIndex]))
                 return false;
         } else {
             // Switch to `Append` mode and remove the existing SVGTransforms starting from `itemIndex`.
@@ -91,7 +143,7 @@ template<typename CharacterType> bool SVGTransformList::parseGeneric(StringParsi
                 resize(itemIndex);
             }
 
-            RefPtr parsedTransform = SVGTransformable::parseTransform(*parsedTransformType, buffer);
+            RefPtr parsedTransform = parseTransform(*parsedTransformType, buffer);
             if (!parsedTransform)
                 return false;
 
@@ -149,4 +201,3 @@ String SVGTransformList::valueAsString() const
 }
 
 }
-
